cmd: distinguish unowned and unmatched files
This commit adds a distinction between files that are not owned by anyone and
files that are not matched by any rule in the CODEOWNERS file.

Example output:

```
Documentation/crdlist.rst                                               (unowned)
Documentation/helm-values.rst                                           (unowned)
proxylib/libcilium.so                                                   (unmatched)
proxylib/libcilium.so.1                                                 (unmatched)
proxylib/libcilium.so.h                                                 (unmatched)
```

To preserve as much of the existing behaviour as possible, the -u flag shows both
unowned and unmatched files, while the -m flag narrows the output down to only
unmatched files.

Signed-off-by: Timo Beckers <timo@incline.eu>